# docs/swagger.yaml
basePath: /api
definitions:
  request.CreateMovieRequest:
    properties:
      description:
        description: |-
          Descripción de la película
          example: "Una aventura épica"
          required: true
        type: string
      duration:
        description: |-
          Duración en minutos
          example: 120
          required: true
        type: integer
      genre:
        description: |-
          Género de la película
          example: "Aventura"
          required: true
        type: string
      image_url:
        description: |-
          URL de la imagen de portada
          example: "https://example.com/image.jpg"
          required: true
        type: string
      stream_url:
        description: |-
          URL del video para streaming
          example: "https://example.com/stream.mp4"
          required: true
        type: string
      title:
        description: |-
          Título de la película
          example: "El Gran Viaje"
          required: true
        type: string
      year:
        description: |-
          Año de lanzamiento de la película
          example: 1980
          required: true
        type: integer
    required:
    - description
    - duration
    - genre
    - image_url
    - stream_url
    - title
    - year
    type: object
...
  request.UpdateMovieRequest:
    properties:
      description:
        description: |-
          Nueva descripción de la película
          example: "Una aventura épica con escenas adicionales"
          required: false
        type: string
      duration:
        description: |-
          Nueva duración en minutos
          example: 150
          required: false
        type: integer
      genre:
        description: |-
          Nuevo género de la película
          example: "Aventura - Edición Especial"
          required: false
        type: string
      image_url:
        description: |-
          Nueva URL de la imagen de portada
          example: "https://example.com/new_image.jpg"
          required: false
        type: string
      stream_url:
        description: |-
          Nueva URL del video para streaming
          example: "https://example.com/new_stream.mp4"
          required: false
        type: string
      title:
        description: |-
          Nuevo título de la película
          example: "El Gran Viaje - Edición Especial"
          required: false
        type: string
      year:
        description: |-
          Nuevo año de lanzamiento
          example: 1981
          required: false
        type: integer
    type: object
...
  response.ErrorResponse:
    properties:
      error:
        description: |-
          Detalle técnico del error (opcional)
          example: "Invalid email format"
        type: string
      message:
        description: |-
          Mensaje descriptivo del error
          example: "Solicitud inválida"
        type: string
      statusCode:
        description: |-
          Código de estado HTTP
          example: 400
        type: integer
    type: object
  response.MovieDetailResponse:
    properties:
      description:
        description: |-
          Descripción extendida de la película
          required: true
          example: La saga de la familia Corleone, liderada por Vito Corleone y posteriormente por su hijo Michael.
        type: string
      genre:
        description: |-
          Género principal de la película
          required: true
          example: Crimen
        type: string
      id:
        description: |-
          ID único de la película en formato string
          required: true
          example: "1"
        type: string
      image_url:
        description: |-
          URL de la imagen en alta calidad
          required: true
          example: https://cloudinary.com/example/padrino_detail.jpg
        type: string
      year:
        description: |-
          Año de lanzamiento
          required: true
          example: 1972
        type: integer
    type: object
  response.MoviePlayerResponse:
    properties:
      id:
        description: |-
          ID único de la película en formato string
          required: true
          example: "1"
        type: string
      stream_url:
        description: |-
          URL del stream de video
          required: true
          example: https://cloudinary.com/example/padrino_stream.m3u8
        type: string
      title:
        description: |-
          Título de la película
          required: true
          example: El Padrino
        type: string
    type: object
  response.MoviePreviewResponse:
    properties:
      id:
        description: |-
          ID único de la película en formato string
          required: true
          example: "1"
        type: string
      image_url:
        description: |-
          URL de la imagen miniatura
          required: true
          example: https://cloudinary.com/example/padrino_thumb.jpg
        type: string
    type: object
  response.MovieResponse:
    properties:
      description:
        description: |-
          Descripción de la película
          required: true
          example: Una épica historia de crimen organizado en la América de los años 40
        type: string
      duration:
        description: |-
          Duración en minutos
          required: true
          example: 175
        type: integer
      genre:
        description: |-
          Género de la película
          required: true
          example: Drama
        type: string
      id:
        description: |-
          ID único de la película
          required: true
          example: 1
        type: integer
      image_url:
        description: |-
          URL de la imagen de portada
          required: true
          example: https://cloudinary.com/example/padrino.jpg
        type: string
      stream_url:
        description: |-
          URL del video para streaming
          required: true
          example: https://cloudinary.com/example/padrino.mp4
        type: string
      title:
        description: |-
          Título de la película
          required: true
          example: El Padrino
        type: string
      year:
        description: |-
          Año de lanzamiento
          required: true
          example: 1972
        type: integer
    type: object
...
  response.SuccessResponse:
    properties:
      data:
        description: Datos adicionales de la operación (opcional)
      message:
        description: |-
          Mensaje descriptivo del resultado
          example: Operación realizada con éxito
        type: string
      success:
        description: |-
          Indica si la operación fue exitosa
          example: true
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  description: API para la plataforma de streaming de películas clásicas.
  title: Vintage Vision API
  version: "1.0"
paths:
  /api/admin/movies:
    post:
      consumes:
      - application/json
      description: Crea una nueva película (recibe las URLs del frontend)
      parameters:
      - description: Datos de la película
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.CreateMovieRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Película creada
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Datos inválidos
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Error al crear
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Crear película
      tags:
      - Movies
  /api/admin/movies/{id}:
    delete:
      description: Elimina una película existente
      parameters:
      - description: ID de la película
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Película eliminada
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: ID inválido
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: No encontrado
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Error al eliminar
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Eliminar película
      tags:
      - Movies
    patch:
      consumes:
      - application/json
      description: Actualiza los datos de una película existente (recibe URLs desde
        el frontend)
      parameters:
      - description: ID de la película
        in: path
        name: id
        required: true
        type: integer
      - description: Datos actualizados
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.UpdateMovieRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Película actualizada
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Datos inválidos
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: No encontrado
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Error al actualizar
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Actualizar película
      tags:
      - Movies
...
  /api/movies:
    get:
      description: Obtiene un listado paginado de películas
      parameters:
      - default: 10
        description: Límite de resultados (default 10)
        in: query
        name: limit
        type: integer
      - default: 0
        description: Desplazamiento (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Listado de películas
          schema:
            items:
              $ref: '#/definitions/response.MovieResponse'
            type: array
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Obtener todas las películas
      tags:
      - Movies
  /api/movies/{id}:
    get:
      description: Obtiene todos los datos de una película
      parameters:
      - description: ID de la película
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Datos completos
          schema:
            $ref: '#/definitions/response.MovieResponse'
        "400":
          description: ID inválido
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: No encontrado
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Obtener película completa
      tags:
      - Movies
  /api/movies/{id}/detail:
    get:
      description: Obtiene información detallada de una película específica
      parameters:
      - description: ID de la película
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Detalles completos
          schema:
            $ref: '#/definitions/response.MovieDetailResponse'
        "400":
          description: ID inválido
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: No encontrado
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Obtener detalles de película
      tags:
      - Movies
  /api/movies/{id}/player:
    get:
      description: Obtiene la información necesaria para reproducir la película
      parameters:
      - description: ID de la película
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Datos para reproductor
          schema:
            $ref: '#/definitions/response.MoviePlayerResponse'
        "400":
          description: ID inválido
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: No encontrado
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Obtener datos para reproductor
      tags:
      - Movies
  /api/movies/by-genre:
    get:
      description: Obtiene películas filtradas por género con paginación
      parameters:
      - description: Género para filtrar
        in: query
        name: genre
        required: true
        type: string
      - default: 1
        description: Página (default 1)
        in: query
        name: page
        type: integer
      - default: 10
        description: Límite por página (default 10)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Resultados paginados
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/response.MovieResponse'
                type: array
              genre:
                type: string
              limit:
                type: integer
              message:
                type: string
              page:
                type: integer
            type: object
        "400":
          description: Parámetros inválidos
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Obtener películas por género
      tags:
      - Movies
  /api/movies/preview:
    get:
      description: Obtiene una vista previa de películas por género
      parameters:
      - description: Género para filtrar
        in: query
        name: genre
        required: true
        type: string
      - default: 1
        description: Página (default 1)
        in: query
        name: page
        type: integer
      - default: 10
        description: Límite por página (default 10)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Vista previa
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/response.MoviePreviewResponse'
                type: array
              genre:
                type: string
              limit:
                type: integer
              message:
                type: string
              page:
                type: integer
            type: object
        "400":
          description: Parámetros inválidos
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Vista previa por género
      tags:
      - Movies
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"